Persist register form
